libft 02

+++ ft_strlen.c
@@ -1,6 +1,8 @@
-int	ft_strlen (const char *s)
+#include "libft.h"
+
+size_t	ft_strlen (const char *s)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (s[len] != '\0')
